+++ document/src/tests/tensor_fieldvalue/partial_remove/partial_remove_test.cpp
@@ -3,7 +3,7 @@
 #include <vespa/eval/eval/simple_value.h>
 #include <vespa/eval/eval/test/tensor_model.hpp>
 #include <vespa/eval/eval/value_codec.h>
-#include <vespa/eval/tensor/partial_update.h>
+#include <vespa/document/update/tensor_partial_update.h>
 #include <vespa/vespalib/util/stringfmt.h>
 #include <vespa/vespalib/gtest/gtest.h>
 #include <optional>
