--- eval/src/vespa/eval/tensor/CMakeLists.txt
@@ -1,5 +1,0 @@
-# Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-vespa_add_library(eval_tensor OBJECT
-    SOURCES
-    partial_update.cpp
-)
